requirements.txt
beautifulsoup4==4.12.3
openai==0.28.0
Requests==2.32.3
streamlit==1.35.0
youtube_transcript_api==0.6.2
python-dotenv
pytube
